requirements.txt
# LLM & Agent Framework
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-experimental>=0.3.0

# Vector Search
opensearch-py>=2.4.0

# Embeddings (Google)
google-generativeai>=0.8.0

# Web Interface
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
